requirements.txt
Django==3.2.*
djangorestframework==3.12.*
psycopg2-binary==2.8.*
python-dotenv==0.17.*
djangorestframework-simplejwt==4.6.*
drf-yasg==1.20.*
gunicorn==20.1.*
whitenoise==5.2.*